requirements.txt
openai==1.50.2
python-dotenv==1.0.1
torch==2.2.2
numpy==1.26.4
requests==2.32.3
vobject==0.9.6.1
httpx==0.27.2
setuptools>=70.0.0,<81
wheel>=0.43.0
sounddevice==0.5.0
webrtcvad==2.0.10
colorama==0.4.6
pvporcupine==3.0.3
pygame==2.6.1
imapclient==3.0.1
pyowm==3.3.0
bs4==0.0.2
pycoingecko==3.1.0
spotipy==2.24.0
PyYAML==6.0.2
